+++ app/components/calendar/CalendarSidebar.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { EventApi, formatDate } from "@fullcalendar/core";
 import { Button } from "@/components/ui/button";
 import { useState } from "react";
-import CalendarSelectModal from "../modals/calendar-select-modal";
+import CalendarSelectModal from "../modals/calendar-view-events-modal";
 
 export default function CalendarSidebar({
   weekendsVisible,
